work on migrating to be an actual library
