+++ pyproject.toml
@@ -1,8 +1,9 @@
 [build-system]
     requires = ["setuptools"]
     build-backend = "setuptools.build_meta"
+
 [tool.setuptools.packages.find]
-    where=["source", "."]
+    where=["source"]
     exclude =["tests"]
